+++ brain/circuit.cpp
@@ -31,23 +31,33 @@ namespace brain
 Circuit::Impl* newImpl(const brion::BlueConfig& config)
 {
     const std::string circuit = config.getCircuitSource().getPath();
+    Circuit::Impl* out;
     if (boost::algorithm::ends_with(circuit, ".mvd2"))
-        return new MVD2(config);
+        out = new MVD2(config);
+    else
+    {
 #ifdef BRAIN_USE_MVD3
-    return new MVD3(config);
+
+        out = new MVD3(config);
 #else
-    throw std::runtime_error("MVD3 support requires CMake 3");
+        throw std::runtime_error("MVD3 not supported");
 #endif
+    }
+    out->_source = brion::URI(config.getSource());
+    return out;
 }
 
 Circuit::Impl* newImpl(const URI& source)
 {
     // Check if sonata
     const auto path = source.getPath();
+    Circuit::Impl* out;
     if (boost::algorithm::ends_with(path, ".json"))
-        return new SonataCircuit(source);
+        out = new SonataCircuit(source);
     else
-        return newImpl(brion::BlueConfig(source.getPath()));
+        out = newImpl(brion::BlueConfig(source.getPath()));
+    out->_source = source;
+    return out;
 }
 
 Circuit::Circuit(const URI& source)
@@ -60,8 +70,16 @@ Circuit::Circuit(const brion::BlueConfig& config)
 {
 }
 
-Circuit::~Circuit()
+Circuit::Circuit(Circuit&& other)
+    : _impl(std::move(other._impl))
 {
+}
+
+Circuit::~Circuit() = default;
+
+const URI& Circuit::getSource() const
+{
+    return _impl->_source;
 }
 
 GIDSet Circuit::getGIDs() const
@@ -205,6 +223,11 @@ Vector3fs Circuit::getPositions(const GIDSet& gids) const
     return _impl->getPositions(gids);
 }
 
+Strings Circuit::getMorphologyNames(const GIDSet& gids) const
+{
+    return _impl->getMorphologyNames(gids);
+}
+
 size_ts Circuit::getMorphologyTypes(const GIDSet& gids) const
 {
     return _impl->getMTypes(gids);
@@ -222,7 +245,7 @@ size_ts Circuit::getElectrophysiologyTypes(const GIDSet& gids) const
 
 Strings Circuit::getElectrophysiologyTypeNames() const
 {
-    return _impl->getElectrophysiologyNames();
+    return _impl->getElectrophysiologyTypeNames();
 }
 
 Matrix4fs Circuit::getTransforms(const GIDSet& gids) const