+++ tests/brain/python/circuit.py
@@ -1,5 +1,5 @@
-# Copyright (c) 2016, EPFL/Blue Brain Project
-#                     Juan Hernando <juan.hernando@epfl.ch>
+# Copyright (c) 2016-2018, EPFL/Blue Brain Project
+#                          Juan Hernando <juan.hernando@epfl.ch>
 #
 # This file is part of Brion <https://github.com/BlueBrain/Brion>
 #
@@ -29,6 +29,7 @@ class TestCircuitOpen(unittest.TestCase):
 
     def test_open(self):
         circuit = brain.Circuit(brain.test.circuit_config)
+        assert(circuit.source() == brain.test.circuit_config)
 
 class TestCircuit(unittest.TestCase):
     def setUp(self):
@@ -45,6 +46,14 @@ class TestCircuit(unittest.TestCase):
 
         assert(len(self.circuit.random_gids(0.1)) == 100)
         assert(len(self.circuit.random_gids(0.1, 'Column')) == 100)
+
+    def test_bad_gids_conversion(self):
+        bad = numpy.array([0.0, 1.0, 2.0])
+        self.assertRaises(ValueError,
+                          lambda: self.circuit.positions(bad))
+        bad = numpy.array(["foo", "bar"])
+        self.assertRaises(ValueError,
+                          lambda: self.circuit.positions(bad))
 
     def test_iteration_order(self):